Minor modification
