+++ docs/src/references/tuning/tune_ewald.rst
@@ -3,5 +3,5 @@ Tune Ewald
 
 .. autofunction:: torchpme.tuning.ewald.tune_ewald
 
-.. autoclass:: torchpme.tuning.error_bounds.EwaldErrorBounds
+.. autoclass:: torchpme.tuning.ewald.EwaldErrorBounds
     :members:
