+++ docs/src/references/tuning/tune_p3m.rst
@@ -3,5 +3,5 @@ Tune P3M
 
 .. autofunction:: torchpme.tuning.p3m.tune_p3m
 
-.. autoclass:: torchpme.tuning.error_bounds.P3MErrorBounds
+.. autoclass:: torchpme.tuning.p3m.P3MErrorBounds
     :members:
