+++ docs/src/references/tuning/tune_pme.rst
@@ -3,5 +3,5 @@ Tune PME
 
 .. autofunction:: torchpme.tuning.pme.tune_pme
 
-.. autoclass:: torchpme.tuning.error_bounds.PMEErrorBounds
+.. autoclass:: torchpme.tuning.pme.PMEErrorBounds
     :members:
