+++ src/torchpme/tuning/ewald.py
@@ -4,7 +4,7 @@ from typing import Optional
 import torch
 
 from ..calculators import EwaldCalculator
-from .tuner import GridSearchTuner
+from .tuner import GridSearchTuner, TuningErrorBounds
 
 TWO_PI = 2 * math.pi
 
@@ -87,6 +87,7 @@ def tune_ewald(
         neighbor_indices=neighbor_indices,
         neighbor_distances=neighbor_distances,
         calculator=EwaldCalculator,
+        error_bounds=EwaldErrorBounds(charges, cell, positions),
         params=params,
     )
     smearing = tuner.estimate_smearing(accuracy)
@@ -99,3 +100,90 @@ def tune_ewald(
         return smearing, params[timings.index(min(timings))]
     # No parameter meets the requirement, return the one with the smallest error
     return smearing, params[errs.index(min(errs))]
+
+
+class EwaldErrorBounds(TuningErrorBounds):
+    r"""
+    Error bounds for :class:`torchpme.calculators.ewald.EwaldCalculator`.
+
+    The error formulas are given `online
+    <https://www2.icp.uni-stuttgart.de/~icp/mediawiki/images/4/4d/Script_Longrange_Interactions.pdf>`_
+    (now not available, need to be updated later). Note the difference notation between
+    the parameters in the reference and ours:
+
+    .. math::
+
+        \alpha &= \left( \sqrt{2}\,\mathrm{smearing} \right)^{-1}
+
+        K &= \frac{2 \pi}{\mathrm{lr\_wavelength}}
+
+        r_c &= \mathrm{cutoff}
+
+    :param charges: atomic charges
+    :param cell: single tensor of shape (3, 3), describing the bounding
+    :param positions: single tensor of shape (``len(charges), 3``) containing the
+        Cartesian positions of all point charges in the system.
+    """
+
+    def __init__(
+        self,
+        charges: torch.Tensor,
+        cell: torch.Tensor,
+        positions: torch.Tensor,
+    ):
+        super().__init__(charges, cell, positions)
+
+        self.volume = torch.abs(torch.det(cell))
+        self.sum_squared_charges = (charges**2).sum()
+        self.prefac = 2 * self.sum_squared_charges / math.sqrt(len(positions))
+        self.cell = cell
+        self.positions = positions
+
+    def err_kspace(
+        self, smearing: torch.Tensor, lr_wavelength: torch.Tensor
+    ) -> torch.Tensor:
+        """
+        The Fourier space error of Ewald.
+
+        :param smearing: see :class:`torchpme.EwaldCalculator` for details
+        :param lr_wavelength: see :class:`torchpme.EwaldCalculator` for details
+        """
+        smearing = torch.as_tensor(smearing)
+        lr_wavelength = torch.as_tensor(lr_wavelength)
+        return (
+            self.prefac**0.5
+            / smearing
+            / torch.sqrt(TWO_PI**2 * self.volume / (lr_wavelength) ** 0.5)
+            * torch.exp(-(TWO_PI**2) * smearing**2 / (lr_wavelength))
+        )
+
+    def err_rspace(self, smearing: torch.Tensor, cutoff: torch.Tensor) -> torch.Tensor:
+        """
+        The real space error of Ewald.
+
+        :param smearing: see :class:`torchpme.EwaldCalculator` for details
+        :param lr_wavelength: see :class:`torchpme.EwaldCalculator` for details
+        """
+        return (
+            self.prefac
+            / torch.sqrt(cutoff * self.volume)
+            * torch.exp(-(cutoff**2) / 2 / smearing**2)
+        )
+
+    def forward(
+        self, smearing: float, lr_wavelength: float, cutoff: float
+    ) -> torch.Tensor:
+        r"""
+        Calculate the error bound of Ewald.
+
+        :param smearing: see :class:`torchpme.EwaldCalculator` for details
+        :param lr_wavelength: see :class:`torchpme.EwaldCalculator` for details
+        :param cutoff: see :class:`torchpme.EwaldCalculator` for details
+        """
+        smearing = torch.as_tensor(smearing)
+        lr_wavelength = torch.as_tensor(lr_wavelength)
+        cutoff = torch.as_tensor(cutoff)
+        return torch.sqrt(
+            self.err_kspace(smearing, lr_wavelength) ** 2
+            + self.err_rspace(smearing, cutoff) ** 2
+        )
