+++ src/torchpme/tuning/pme.py
@@ -5,7 +5,7 @@ from typing import Optional
 import torch
 
 from ..calculators import PMECalculator
-from .tuner import GridSearchTuner
+from .tuner import GridSearchTuner, TuningErrorBounds
 
 TWO_PI = 2 * math.pi
 
@@ -98,6 +98,7 @@ def tune_pme(
         neighbor_indices=neighbor_indices,
         neighbor_distances=neighbor_distances,
         calculator=PMECalculator,
+        error_bounds=PMEErrorBounds,
         params=params,
     )
     smearing = tuner.estimate_smearing(accuracy)
@@ -110,3 +111,110 @@ def tune_pme(
         return smearing, params[timings.index(min(timings))]
     # No parameter meets the requirement, return the one with the smallest error
     return smearing, params[errs.index(min(errs))]
+
+
+class PMEErrorBounds(TuningErrorBounds):
+    r"""
+    Error bounds for :class:`torchpme.PMECalculator`.
+    For the error formulas are given `elsewhere <https://doi.org/10.1063/1.470043>`_.
+    Note the difference notation between the parameters in the reference and ours:
+
+    .. math::
+
+        \alpha = \left(\sqrt{2}\,\mathrm{smearing} \right)^{-1}
+
+    :param charges: atomic charges
+    :param cell: single tensor of shape (3, 3), describing the bounding
+    :param positions: single tensor of shape (``len(charges), 3``) containing the
+        Cartesian positions of all point charges in the system.
+    """
+
+    def __init__(
+        self, charges: torch.Tensor, cell: torch.Tensor, positions: torch.Tensor
+    ):
+        super().__init__(charges, cell, positions)
+
+        self.volume = torch.abs(torch.det(cell))
+        self.sum_squared_charges = (charges**2).sum()
+        self.prefac = 2 * self.sum_squared_charges / math.sqrt(len(positions))
+        self.cell_dimensions = torch.linalg.norm(cell, dim=1)
+
+    def err_kspace(
+        self,
+        smearing: torch.Tensor,
+        mesh_spacing: torch.Tensor,
+        interpolation_nodes: torch.Tensor,
+    ) -> torch.Tensor:
+        """
+        The Fourier space error of PME.
+
+        :param smearing: see :class:`torchpme.PMECalculator` for details
+        :param mesh_spacing: see :class:`torchpme.PMECalculator` for details
+        :param interpolation_nodes: see :class:`torchpme.PMECalculator` for details
+        """
+        actual_spacing = self.cell_dimensions / (
+            2 * self.cell_dimensions / mesh_spacing + 1
+        )
+        h = torch.prod(actual_spacing) ** (1 / 3)
+        i_n_factorial = torch.exp(torch.lgamma(interpolation_nodes + 1))
+        RMS_phi = [None, None, 0.246, 0.404, 0.950, 2.51, 8.42]
+
+        return (
+            self.prefac
+            * torch.pi**0.25
+            * (6 * (1 / 2**0.5 / smearing) / (2 * interpolation_nodes + 1)) ** 0.5
+            / self.volume ** (2 / 3)
+            * (2**0.5 / smearing * h) ** interpolation_nodes
+            / i_n_factorial
+            * torch.exp(
+                interpolation_nodes * (torch.log(interpolation_nodes / 2) - 1) / 2
+            )
+            * RMS_phi[interpolation_nodes - 1]
+        )
+
+    def err_rspace(self, smearing: torch.Tensor, cutoff: torch.Tensor) -> torch.Tensor:
+        """
+        The real space error of PME.
+
+        :param smearing: see :class:`torchpme.PMECalculator` for details
+        :param cutoff: see :class:`torchpme.PMECalculator` for details
+        """
+        smearing = torch.as_tensor(smearing)
+        cutoff = torch.as_tensor(cutoff)
+
+        return (
+            self.prefac
+            / torch.sqrt(cutoff * self.volume)
+            * torch.exp(-(cutoff**2) / 2 / smearing**2)
+        )
+
+    def error(
+        self,
+        cutoff: float,
+        smearing: float,
+        mesh_spacing: float,
+        interpolation_nodes: float,
+    ) -> torch.Tensor:
+        r"""
+        Calculate the error bound of PME.
+
+        :param smearing: if its value is given, it will not be tuned, see
+            :class:`torchpme.PMECalculator` for details
+        :param mesh_spacing: if its value is given, it will not be tuned, see
+            :class:`torchpme.PMECalculator` for details
+        :param cutoff: if its value is given, it will not be tuned, see
+            :class:`torchpme.PMECalculator` for details
+        :param interpolation_nodes: The number ``n`` of nodes used in the interpolation
+            per coordinate axis. The total number of interpolation nodes in 3D will be
+            ``n^3``. In general, for ``n`` nodes, the interpolation will be performed by
+            piecewise polynomials of degree ``n - 1`` (e.g. ``n = 4`` for cubic
+            interpolation). Only the values ``3, 4, 5, 6, 7`` are supported.
+        """
+        smearing = torch.as_tensor(smearing)
+        mesh_spacing = torch.as_tensor(mesh_spacing)
+        cutoff = torch.as_tensor(cutoff)
+        interpolation_nodes = torch.as_tensor(interpolation_nodes)
+        return torch.sqrt(
+            self.err_rspace(smearing, cutoff) ** 2
+            + self.err_kspace(smearing, mesh_spacing, interpolation_nodes) ** 2
+        )
